GH #47: Manage bareword filehandles in runtime:
(What this solution does appears further down after "Context.")

Context
-------

When Test::MockFile loads, it overrides a set of core functions like
open(). It uses prototypes that *should* match the same protoypes as
the core functions. (`prototype("CORE::open")` is example syntax for
retrieving such prototypes.

However, the prototypes are not actually the same as the parsing perl
does. It's just the closest approximation that can be expressed in
prototypes. that's usually not a problem, except when dealing with
bareword filehandles.

There are two cases:

1. strict refs failures (runtime)
2. strict subs failures (compile-time)

Let's explore:

1. The strict refs failures are detected during runtime. A good example
of it is GH #47:

    use Test::MockFile (); # override open
    use File::Find     (); # will use open in runtime, so no error yet

    # Call opendir() which was overridden by Test::MockFile
    # using a bareword, which causes a strict refs failure
    File::Find::find( sub {1}, '/' );

    print "Never reach here due to runtime error";

    # This is what's actually in File::Find::find():
    #   opendir DIR, ...

Technically, this *isn't* actually a stricture problem because `perl`
will parse it correctly, but *we* cannot parse it correctly with the
prototypes we use - or any prototype, really.

(It's important to note that there's another side-effect that happens here
which is described below under 'What the actual issue is.')

2. The strict subs are triggered for the same issue, but when it happens
during compile time. For example:

    # Override open()
    use Test::MockFile ();

    # Calls open() with bareword
    use Term::Table::Util ();

    print "Never reach here due to compile-time error";

    # This is what's actually in Term::Table:
    #   BEGIN { open( $IO, '>&', STDOUT or die... }

Here's a fun fact:

    # This works just fine
    open my $fh, '>&', STDOUT;

    # This fails with strict
    open my $fh, '>', STDOUT;

This shows that the parsing rules are not consistent and cannot be
expressed with prototypes. That is, they should both fail on strict but
`perl` will understand the `STDOUT` in the first version.

There's nothing we can do about this type of error. This is because of the
actual underlying issue described below. There's nothing that we can do
about it, other than loading `Test::MockFile` at the end.

What the actual issue is
------------------------

The following order of events happens:

* `Test::MockFile` is loaded
* `Test::MockFile` overrides core functions like `open`, `opendir`, etc.
* A module, like `File::Find`, is loaded
* `File::Find::find()` will call `opendir DIR` which is a bareword filehandle
* Our `opendir()` function will be called, but because of the parsing rules,
  the argument we will receive is the string `DIR`.
* We will send this string to `CORE::opendir()` which will get mad because it
  cannot use that string. It mentions the user's code as the caller because we
  use `goto &CORE::open` which will remove *us* from the calling stack.

This means our issue is that our code will receive the bareword (like `STDOUT`)
as a string (`"STDOUT"`) and then we pass *that* to the core `open()` function.

This is why any function call that is parsed correctly (and sends us the bareword
as a string) can be fixed and any function call that now cannot be parsed
correctly (because our prototypes are not enough) cannot be fixed.

Solution
--------

Barewords are essentially evaluated as strings and then looked up
as a typeglob with an 'IO' entry.

What we do is try to figure out if this is a bareword filehandle by
looking up a typeglob with the string representing the variable name
and if that works, we use that instead.

This is slightly tricky and I wish I could make it even more specific
by checking for an IO entry. However, when this is mocked, it's a bit
trickier and the `IO` slot will not be available.