GH #34: Support open() with barewords:
The last change to support barewords did not include `open()` by accident,
but it wouldn't have helped anyway because of how `open()` operates with
barewords calling them as `<FH>` and how `Test::MockFile` attempts to write
over `$_[0]` which is what caused the original error expressed in the ticket
(modification of read-only value).

So here's what we're doing differently:

* We added bareword upgrading to `open()`.

* We try to detect barewords using the READONLY flag which is what they will
  have.

* We retrieve the value of the bareword slot using the right caller, which
  is probably only relevant for barewords that already existed and have a
  value we want. I'm not sure what those situations will be, but okay.

* The bareword upgrading also now returns information on whether this is
  likely to be a bareword.

* This information is only used with `open()` at the moment to do two things:
  * Override `@_`, not `$_[0]`
  * Override the bareword slot using the caller information

The likely bareword information is removed in places where it's not used.

It's certainly more verbose but I think it's still for the better.