Do not get confused by inner directory files:
We would grab all files in the directory recursively, but directory
listing should only show the directories inside it, not the files
within those directories.

In other words:

    foo/
        bar/
            baz.txt

`readdir($path_to_foo)` should show `bar`, not `bar/baz.txt`.

Resolves GH #118

Unfortunately, because of our naive structure types (an array of string),
we may find both a directory and a file within that directory, so we also
need to uniq'ify the results.