Do not corrupt handles:
When `_upgrade_barewords()` is called, it checks the reference and
decides to skip otherwise. The problem is that we always reset `@_`,
even when it returns. This messes up `@_` and in cases where we have
a handle on the stack, it will not be changed to an `IO::File` instance.

Instead, we do the checks outside like we do with `defined`.

There doesn't seem to be a simple way to change the caller's caller's
stack in pure Perl.