Introduce new_dir() to allow older dir() syntax:
With this method, you can now both create a directory and set the
mode for it at the same time, as the old `dir()` keyword allowed.

    my $dir = Test::MockFile->new_dir( '/path', { 'mode' => 0644 } );

The one thing that is not supported (and shouldn't be until a proper
interface for it is defined) is setting the contents of that directory.

    # This will crash
    my $dir = Test::MockFile->new_dir( '/path', [...], ... ); # CRASH

    # This will also crash
    my $dir = Test::MockFile->new_dir('/path');
    $dir->contents( ... ); # CRASH

However, you can still set the contents of a file using `contents()`:

    $file->contents('new content');

ALSO:

To do this, I had to move all the keywords to functions I'm calling.
This is because `CORE::GLOBAL::mkdir()` failed on 5.16 and below.